update:添加UI界面、跳转进度逻辑，处理音视频同步问题

+++ JR.VPlayer/VideoPacket.cs
@@ -12,10 +12,10 @@ namespace JR.VPlayer
         public readonly AVPixelFormat Format;
         public readonly int Width;
         public readonly int Height;
-        public readonly int Second;
+        public readonly long Second;
         public byte[] Data;
 
-        public VideoPacket(AVPixelFormat format, int width, int height, int second,byte[] data)
+        public VideoPacket(AVPixelFormat format, int width, int height, long second,byte[] data)
         {
             this.Format = format;
             this.Width = width;
